# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: BOOK 조회용 API
  description: Spring-OAS-Docs API 샘플
  version: "1.0"
servers:
- url: http://localhost:8080/
tags: []
paths:
  /books:
    get:
      tags:
      - books
      operationId: 책 전체 조회
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/books413989036'
              examples:
                책 전체 조회:
                  value: 
                    [ {
                      "id" : 1,
                      "title" : "호부호형",
                      "author" : "홍길동",
                      "uuid" : "550e8400-e29b-41d4-a716-446655440000",
                      "description" : "의적 이야기",
                      "createdAt" : "2022-09-25 08:51:51",
                      "updatedAt" : "2022-09-25 08:51:51"
                    }, {
                      "id" : 2,
                      "title" : "도깨비",
                      "author" : "이신",
                      "uuid" : "912a9vja-c2js-1sca-g6hy-snx1dms34851",
                      "description" : null,
                      "createdAt" : "2022-09-25 08:51:51",
                      "updatedAt" : "2022-09-25 08:51:51"
                    }, {
                      "id" : 3,
                      "title" : "제주도푸른밤",
                      "author" : "성식영",
                      "uuid" : "a9xa8241-s831-jja9-vkd1-8912a3dn8801",
                      "description" : "노래와 음식",
                      "createdAt" : "2022-09-25 08:51:51",
                      "updatedAt" : "2022-09-25 08:51:51"
                    }, {
                      "id" : 4,
                      "title" : "반려동물 교육",
                      "author" : "강동생",
                      "uuid" : "c82qjf1a-0023-9gma-9244-bbk19snd184b",
                      "description" : null,
                      "createdAt" : "2022-09-25 08:51:51",
                      "updatedAt" : "2022-09-25 08:51:51"
                    }, {
                      "id" : 5,
                      "title" : "워킹 디벨로퍼",
                      "author" : "WKingDev",
                      "uuid" : "018nv881-10fm-84nf-82mz-001mf98j02j9",
                      "description" : "개발자",
                      "createdAt" : "2022-09-25 08:51:51",
                      "updatedAt" : "2022-09-25 08:51:51"
                    } ]
  /books/{id}:
    get:
      tags:
      - books
      operationId: 책 단일 조회
      parameters:
      - name: id
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/books-id594339163'
              examples:
                책 단일 조회:
                  value: 
                    {
                      "id" : 5,
                      "title" : "워킹 디벨로퍼",
                      "author" : "WKingDev",
                      "uuid" : "018nv881-10fm-84nf-82mz-001mf98j02j9",
                      "description" : "개발자",
                      "createdAt" : "2022-09-25 08:51:51",
                      "updatedAt" : "2022-09-25 08:51:51"
                    }
components:
  schemas:
    books413989036:
      type: array
      items:
        type: object
        properties:
          createdAt:
            type: string
            description: 생성 일시
          author:
            type: string
            description: 작가
          description:
            type: string
            description: 설명
          id:
            type: number
            description: 책 아이디
          title:
            type: string
            description: 제목
          uuid:
            type: string
            description: 일련 번호
          updatedAt:
            type: string
            description: 수정 일시
    books-id594339163:
      type: object
      properties:
        createdAt:
          type: string
          description: 생성 일시
        author:
          type: string
          description: 작가
        description:
          type: string
          description: 설명
        id:
          type: number
          description: 책 아이디
        title:
          type: string
          description: 제목
        uuid:
          type: string
          description: 일련 번호
        updatedAt:
          type: string
          description: 수정 일시